+++ NEWS.md
@@ -5,6 +5,9 @@
 # Version 2.3.0
 - New control parameter 'fillMissingSpParams'
 - New species categorical params (for inbuilt imputation) 'LeafShape' and 'LeafSize'
+- Parameter 'Flammability' index eliminated (non-meansurable property of flammability).
+- Parameter 'LeafLitterFuelType' eliminated because it is derived from 'LeafShape' and 'LeafSize'.
+- Defaults added for 'heatContent' and 'SAV' according to 'LeafShape' and 'LeafSize'
 
 # Version 2.2.3
 - Functions soilgridsParams() and forest_map*Tables() moved to package 'medfateutils' available at GitHub (emf-creaf/medfateutils).